Debugging

+++ src/ECS/Components/PositionComponent.hpp
@@ -2,10 +2,10 @@
 
 
 struct PositionComponent {
-	float x;
-	float y;
+	int x;
+	int y;
 
-	explicit PositionComponent(float x = 0, float y = 0) :
+	explicit PositionComponent(int x = 0, int y = 0) :
 	  x(x), y(y) {
 	}
 };
